refact: seperate test handler
